requirements.txt
flask
pytest
flake8
pandas
google-cloud-secret-manager
google-cloud-storage
boto3==1.34.59
google-auth
